Split local placer modules
